* many changes for correct utf-8 support

+++ antargis/trunk/src/mesh_optimizer.cc
@@ -36,7 +36,7 @@ MeshOptimizer::MeshOptimizer()
 }
 void MeshOptimizer::add(const MeshVertex &v)
 {
-  int index=mVertices.size();
+  size_t index=mVertices.size();
   std::map<MeshVertex,int>::iterator i=mMap.find(v);
   if(i!=mMap.end())
     {
